Gh 381: retrain classification and sentiment models with latest version of scikit-learn (#383)

+++ underthesea/model_fetcher.py
@@ -18,11 +18,13 @@ with open(join(CD, "models.yaml")) as f:
 class UTSModel(Enum):
     tc_general = "TC_GENERAL"
     tc_bank = "TC_BANK"
-
+    tc_general_v131 = "TC_GENERAL_V131"
+    tc_bank_v131 = "TC_BANK_V131"
     sa_general = "SA_GENERAL"
     sa_bank = "SA_BANK"
+    sa_bank_v131 = "SA_BANK_V131"
 
-
+# flake8: noqa: C901
 class ModelFetcher:
 
     @staticmethod
@@ -48,6 +50,33 @@ class ModelFetcher:
             os.rename(
                 Path(CACHE_ROOT) / cache_dir / "tc_svm_vntc_20190607",
                 Path(CACHE_ROOT) / cache_dir / "TC_GENERAL",
+            )
+            os.remove(model_path)
+
+        if model_name == "TC_GENERAL_V131":
+            url = "https://github.com/undertheseanlp/playground/releases/download/1.3.x/tc_svm_vntc_20201228.zip"
+            cached_path(url, cache_dir=cache_dir)
+            model_path = Path(CACHE_ROOT) / cache_dir / "tc_svm_vntc_20201228.zip"
+            cache_folder = Path(CACHE_ROOT) / cache_dir
+            zip = zipfile.ZipFile(model_path)
+            zip.extractall(cache_folder)
+            os.rename(
+                Path(CACHE_ROOT) / cache_dir / "tc_svm_vntc_20201228",
+                Path(CACHE_ROOT) / cache_dir / "TC_GENERAL_V131",
+            )
+            os.remove(model_path)
+
+        if model_name == "TC_BANK_V131":
+            model_id = "tc_svm_ubc-1_20210107"
+            url = f"https://github.com/undertheseanlp/playground/releases/download/1.3.x/{model_id}.zip"
+            cached_path(url, cache_dir=cache_dir)
+            model_path = Path(CACHE_ROOT) / cache_dir / f"{model_id}.zip"
+            cache_folder = Path(CACHE_ROOT) / cache_dir
+            zip = zipfile.ZipFile(model_path)
+            zip.extractall(cache_folder)
+            os.rename(
+                Path(CACHE_ROOT) / cache_dir / model_id,
+                Path(CACHE_ROOT) / cache_dir / model_name,
             )
             os.remove(model_path)
 
@@ -77,6 +106,20 @@ class ModelFetcher:
             )
             os.remove(model_path)
 
+        if model_name == "SA_GENERAL_V131":
+            model_id = "sa_svm_vlsp2016-sa_20210107"
+            url = f"https://github.com/undertheseanlp/playground/releases/download/1.3.x/{model_id}.zip"
+            cached_path(url, cache_dir=cache_dir)
+            model_path = Path(CACHE_ROOT) / cache_dir / f"{model_id}.zip"
+            cache_folder = Path(CACHE_ROOT) / cache_dir
+            zip = zipfile.ZipFile(model_path)
+            zip.extractall(cache_folder)
+            os.rename(
+                Path(CACHE_ROOT) / cache_dir / model_id,
+                Path(CACHE_ROOT) / cache_dir / model_name,
+            )
+            os.remove(model_path)
+
         if model_name == "SA_BANK":
             url = "https://www.dropbox.com/s/yo6sf6ofpdb3hlh/sa_svm_uts2017_bank_20190611.zip?dl=1"
             cached_path(url, cache_dir=cache_dir)
@@ -87,6 +130,20 @@ class ModelFetcher:
             os.rename(
                 Path(CACHE_ROOT) / cache_dir / "sa_svm_uts2017_bank_20190611",
                 Path(CACHE_ROOT) / cache_dir / "SA_BANK",
+            )
+            os.remove(model_path)
+
+        if model_name == "SA_BANK_V131":
+            model_id = "sa_svm_ubs-1_20210107"
+            url = f"https://github.com/undertheseanlp/playground/releases/download/1.3.x/{model_id}.zip"
+            cached_path(url, cache_dir=cache_dir)
+            model_path = Path(CACHE_ROOT) / cache_dir / f"{model_id}.zip"
+            cache_folder = Path(CACHE_ROOT) / cache_dir
+            zip = zipfile.ZipFile(model_path)
+            zip.extractall(cache_folder)
+            os.rename(
+                Path(CACHE_ROOT) / cache_dir / model_id,
+                Path(CACHE_ROOT) / cache_dir / model_name,
             )
             os.remove(model_path)
 
@@ -135,3 +192,4 @@ class ModelFetcher:
 
         if model == UTSModel.sa_bank:
             return Path(CACHE_ROOT) / "models" / "SA_BANK"
+        return Path(CACHE_ROOT) / "models" / model
